backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pandas>=2.1.0
pyarrow>=14.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
plotly>=5.23.0
anthropic>=0.40.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
